+++ src/editor/core/plugins/anchors/index.js
@@ -6,10 +6,14 @@
 
 import anchor_plugin from "markdown-it-anchor";
 
-const position = {
-    false: 'push',
-    true: 'unshift'
-};
+const copyHrefToClipboard = (target, context) => {
+    const href = target.href || target.parentElement.href;
+    if (href) {
+        navigator.clipboard.writeText(href).then(r => {
+            humhub.modules.ui.status.success(context.translate('Copied'));
+        });
+    }
+}
 
 const anchors = {
     id: 'anchor',
@@ -17,16 +21,16 @@ const anchors = {
     init: (context, isEdit) => {
         if (!isEdit) {
             context.editor.$.on('mouseenter', ':header', function () {
-                $(this).find('.header-anchor').show();
+                $(this).find('.header-anchor').show().on('click', (e) => copyHrefToClipboard(e.target, context));
             }).on('mouseleave', ':header', function () {
-                $(this).find('.header-anchor').hide();
+                $(this).find('.header-anchor').hide().off('click');
             });
         }
     },
     registerMarkdownIt: (markdownIt) => {
         const anchorOptions = {
             permalink: anchor_plugin.permalink.linkInsideHeader({
-                symbol: '¶',
+                symbol: '<i class="fa fa-chain"></i>',
                 placement: 'after',
                 ariaHidden: true,
                 renderAttrs: (slug, state) => {