+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 2.0.2 (July 28, 2023)
 ---------------------
 - Fix #113: Header anchors lost
+- Enh #115: Headline Anchor - Change Icon
 
 
 2.0.0